refactor: split large basic.rs test file into smaller, organized modules
- basic_expressions.rs: basic SELECT expressions, literals, arithmetic, wildcards
- all_distinct.rs: SELECT ALL syntax and DISTINCT verification tests
- derived_columns.rs: derived column lists for wildcards
- backtick_identifiers.rs: MySQL-style backtick identifier tests

This reduces the largest test file from 685 lines to ~150-200 lines each,
making tests more maintainable and easier to navigate.

Amp-Thread-ID: https://ampcode.com/threads/T-e168ad32-d9ab-4695-a238-1f5d61214711
