+++ crates/parser/src/tests/select/mod.rs
@@ -1,4 +1,7 @@
-mod basic;
+mod all_distinct;
+mod backtick_identifiers;
+mod basic_expressions;
 mod concat;
+mod derived_columns;
 mod filters;
 mod position;
